Crate servo_hyper_serde

This crate provides wrappers and convenience functions to make Hyper and Serde work hand in hand.

The supported types are:

cookie::Cookie

headers_ext::ContentType

hyper::header::Headers

hyper::StatusCode

hyper::Method

hyper::Uri

mime::Mime

How do I use a data type with a HeaderMap member with Serde?

Use the serde attributes deserialize_with and serialize_with.

#[derive(serde::Deserialize, serde::Serialize)]
struct MyStruct {
    #[serde(deserialize_with = "servo_hyper_serde::deserialize",
            serialize_with = "servo_hyper_serde::serialize")]
    headers: http::header::HeaderMap,
}

How do I encode a HeaderMap value with serde_json::to_string?

Use the Ser wrapper.

use servo_hyper_serde::Ser;
let headers = http::header::HeaderMap::new();
serde_json::to_string(&Ser::new(&headers));

How do I decode a Method value with serde_json::from_str?

Use the De wrapper.

use servo_hyper_serde::De;
use hyper::Method;
serde_json::from_str::<De<Method>>("\"PUT\"").map(De::into_inner);

How do I send Cookie values as part of an IPC channel?

Use the Serde wrapper. It implements Deref and DerefMut for convenience.

use servo_hyper_serde::Serde;
use cookie::Cookie;
ipc_channel::ipc::channel::<Serde<Cookie>>();

Structs

De

A wrapper to deserialize Hyper types.

Ser

A wrapper to serialize Hyper types.

Serde

A convenience wrapper to be used as a type parameter, for example when a Vec<T> need to be passed to serde.

StatusVisitor 🔒

Functions

deserialize

Deserialises a T value with a given deserializer.

serialize

Serialises value with a given serializer.

serialize_pretty

Serialises value with a given serializer in a pretty way.